Fixed Bug #1805 and improved memory needs again
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@1781 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribusstructs.h
@@ -24,17 +24,17 @@ struct ScText
 {
 	bool cselect;
 	int csize;
-	int cshade;
-	int cshade2;
-	int cstyle;
-	int cab;
-	int cscale;
-	double cextra;
-	double xp;
-	double yp;
-	double PtransX;
-	double PtransY;
-	double PRot;
+	short cshade;
+	short cshade2;
+	short cstyle;
+	short cab;
+	short cscale;
+	float cextra;
+	float xp;
+	float yp;
+	float PtransX;
+	float PtransY;
+	float PRot;
 	QString cfont;
 	QString ccolor;
 	QString cstroke;
